moved save dialogue to better place
